Stop importing form_setup selectors outside setup
Common selectors are in form/selectors now.
