+++ indico/modules/events/registration/client/js/form/fields/BooleanInput.jsx
@@ -13,7 +13,7 @@ import {Button, Label} from 'semantic-ui-react';
 import {FinalDropdown, FinalField, FinalInput, validators as v} from 'indico/react/forms';
 import {Translate} from 'indico/react/i18n';
 
-import {getCurrency} from '../../form_setup/selectors';
+import {getCurrency} from '../../form/selectors';
 
 import {PlacesLeft} from './PlacesLeftLabel';
