+++ indico/modules/events/registration/client/js/form/fields/MultiChoiceInput.jsx
@@ -15,7 +15,7 @@ import {Checkbox, Dropdown, Label} from 'semantic-ui-react';
 import {FinalCheckbox, FinalField} from 'indico/react/forms';
 import {Translate} from 'indico/react/i18n';
 
-import {getCurrency} from '../../form_setup/selectors';
+import {getCurrency} from '../../form/selectors';
 
 import {Choices, choiceShape} from './ChoicesSetup';
 import {PlacesLeft} from './PlacesLeftLabel';
