+++ indico/modules/events/registration/client/js/form_setup/ItemSettingsModal.jsx
@@ -24,10 +24,11 @@ import {FinalModalForm} from 'indico/react/forms/final-form';
 import {Translate, Param} from 'indico/react/i18n';
 
 import {fieldRegistry} from '../form/fields/registry';
+import {getStaticData} from '../form/selectors';
 
 import * as actions from './actions';
 import ItemTypeDropdown from './ItemTypeDropdown';
-import {getStaticData, getItemById} from './selectors';
+import {getItemById} from './selectors';
 
 export default function ItemSettingsModal({id, sectionId, defaultNewItemType, onClose}) {
   const dispatch = useDispatch();
