+++ indico/modules/events/registration/client/js/form_setup/selectors.js
@@ -8,51 +8,14 @@
 import _ from 'lodash';
 import {createSelector} from 'reselect';
 
-export const isUILocked = state => state.uiLocked;
-export const getStaticData = state => state.staticData;
-const getFlatSections = state => state.sections;
-const getItems = state => state.items;
+import {getStaticData, getFlatSections, getItems} from '../form/selectors';
 
-/** Get the currency used by the registration form. */
-export const getCurrency = createSelector(
-  getStaticData,
-  staticData => staticData.currency
-);
+export const isUILocked = state => state.uiLocked;
 
 /** Get a sorted list of disabled sections. */
 export const getDisabledSections = createSelector(
   getFlatSections,
   sections => _.sortBy(Object.values(sections).filter(s => !s.enabled), ['position', 'id'])
-);
-
-/** Get a sorted list of enabled top-level sections. */
-const getSections = createSelector(
-  getFlatSections,
-  sections => _.sortBy(Object.values(sections).filter(s => s.enabled), ['position', 'id'])
-);
-
-/** Get a mapping from section IDs to sorted item lists. */
-const getSectionItemMapping = createSelector(
-  getFlatSections,
-  getItems,
-  (sections, fields) => {
-    const mapping = new Map(Object.values(sections).map(s => [s.id, []]));
-    _.sortBy(Object.values(fields), ['position', 'id']).forEach(field => {
-      mapping.get(field.sectionId).push(field);
-    });
-    return mapping;
-  }
-);
-
-/** Get the nested form structure of sections and their items. */
-export const getNestedSections = createSelector(
-  getSections,
-  getSectionItemMapping,
-  (sortedSections, sectionFields) =>
-    sortedSections.map(section => ({
-      ...section,
-      items: sectionFields.get(section.id),
-    }))
 );
 
 /** Get a section by its ID. */
