+++ indico/modules/events/registration/client/js/form_submission/selectors.js
@@ -7,11 +7,7 @@
 
 import {createSelector} from 'reselect';
 
-import {getStaticData} from '../form_setup/selectors';
-
-// TODO: move common selectors (ie most of the section and currency stuff) into ../form/selectors.js
-export {getNestedSections} from '../form_setup/selectors';
-export {getStaticData} from '../form_setup/selectors';
+import {getStaticData} from '../form/selectors';
 
 export const getUserInfo = createSelector(
   getStaticData,
